+++ src/structure/intpoly.cpp
@@ -1,6 +1,6 @@
 #include "structure/intpoly.hpp"
 
-IntPoly::IntPoly(size_t N) : PolyBase<uint32_t>(N) {};
+IntPoly::IntPoly(uint32_t N) : PolyBase<uint32_t>(N) {};
 
 IntPoly::IntPoly(const std::vector<uint32_t> &coeffs) : PolyBase<uint32_t>(coeffs) {};
